refactor: movido los dto a carpeta swagger
