+++ src/swagger/dto/boe.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { Boe } from '../models/boe.model';
+import { Boe } from '../../compartido/models/boe.model';
 export class BoeDTO implements Boe {
   @ApiProperty({ example: ['BOE-B-AAAA-XXX1', 'BOE-B-AAAA-XXX2'], description: 'Colección de ID de contratos' })
   idContratos: string[];
